Remove TypeRestrictions interface
