+++ generator/src/main/java/com/scottlogic/deg/generator/generation/StandardFieldValueSourceEvaluator.java
@@ -47,7 +47,7 @@ public class StandardFieldValueSourceEvaluator implements FieldValueSourceEvalua
 
         TypeRestrictions typeRestrictions = fieldSpec.getTypeRestrictions() != null
             ? fieldSpec.getTypeRestrictions()
-            : DataTypeRestrictions.ALL_TYPES_PERMITTED;
+            : TypeRestrictions.ALL_TYPES_PERMITTED;
 
         if (typeRestrictions.isTypeAllowed(IsOfTypeConstraint.Types.NUMERIC)) {
             validSources.add(getNumericSource(fieldSpec));
