+++ generator/src/main/java/com/scottlogic/deg/generator/restrictions/TypeRestrictions.java
@@ -17,15 +17,77 @@
 package com.scottlogic.deg.generator.restrictions;
 
 import com.scottlogic.deg.common.profile.constraints.atomic.IsOfTypeConstraint;
+import com.scottlogic.deg.generator.utils.SetUtils;
 
-import java.util.Set;
+import java.util.*;
 
-public interface TypeRestrictions extends Restrictions {
-    TypeRestrictions except(IsOfTypeConstraint.Types... types);
+public class TypeRestrictions implements Restrictions {
 
-    boolean isTypeAllowed(IsOfTypeConstraint.Types type);
+    public static final TypeRestrictions ALL_TYPES_PERMITTED = new TypeRestrictions(Arrays.asList(IsOfTypeConstraint.Types.values()));
 
-    TypeRestrictions intersect(TypeRestrictions other);
+    public TypeRestrictions(Collection<IsOfTypeConstraint.Types> allowedTypes) {
+        if (allowedTypes.isEmpty())
+            throw new UnsupportedOperationException("Cannot have a type restriction with no types");
 
-    Set<IsOfTypeConstraint.Types> getAllowedTypes();
+        this.allowedTypes = new HashSet<>(allowedTypes);
+    }
+
+    public static TypeRestrictions createFromWhiteList(IsOfTypeConstraint.Types... types) {
+        return new TypeRestrictions(Arrays.asList(types));
+    }
+
+    public TypeRestrictions except(IsOfTypeConstraint.Types... types) {
+        if (types.length == 0)
+            return this;
+
+        ArrayList<IsOfTypeConstraint.Types> allowedTypes = new ArrayList<>(this.allowedTypes);
+        allowedTypes.removeAll(Arrays.asList(types));
+
+        return new TypeRestrictions(allowedTypes);
+    }
+
+    private final Set<IsOfTypeConstraint.Types> allowedTypes;
+
+    public boolean isTypeAllowed(IsOfTypeConstraint.Types type){
+        return allowedTypes.contains(type);
+    }
+
+    public String toString() {
+        if (allowedTypes.size() == 1)
+            return String.format("Type = %s", allowedTypes.toArray()[0]);
+
+        return String.format(
+                "Types: %s",
+                Objects.toString(allowedTypes));
+    }
+
+    public TypeRestrictions intersect(TypeRestrictions other) {
+        if (other == ALL_TYPES_PERMITTED)
+            return this;
+
+        Set<IsOfTypeConstraint.Types> intersection = SetUtils.intersect(allowedTypes, other.getAllowedTypes());
+
+        if (intersection.isEmpty())
+            return null;
+
+        return new TypeRestrictions(intersection);
+    }
+
+    public Set<IsOfTypeConstraint.Types> getAllowedTypes() {
+        return allowedTypes;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        TypeRestrictions that = (TypeRestrictions) o;
+        return Objects.equals(allowedTypes, that.allowedTypes);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(allowedTypes);
+    }
 }
+
