+++ generator/src/test/java/com/scottlogic/deg/generator/generation/StandardFieldValueSourceEvaluatorTests.java
@@ -85,7 +85,7 @@ public class StandardFieldValueSourceEvaluatorTests {
             min = new NumericLimit<>(new BigDecimal(10), false);
             max = new NumericLimit<>(new BigDecimal(30), false);
         }};
-        TypeRestrictions typeRestrictions = new DataTypeRestrictions(Collections.singletonList(
+        TypeRestrictions typeRestrictions = new TypeRestrictions(Collections.singletonList(
             IsOfTypeConstraint.Types.NUMERIC
         ));
         FieldSpec fieldSpecWithTypedNumericRestrictionsAndNullNotDisallowed = FieldSpec.Empty
@@ -101,7 +101,7 @@ public class StandardFieldValueSourceEvaluatorTests {
     public void shouldReturnNullSourceLastWithTypedStringRestrictionsAndNullNotDisallowed() {
         StandardFieldValueSourceEvaluator evaluator = new StandardFieldValueSourceEvaluator();
         StringRestrictions stringRestrictions = matchesRegex("/[ab]{2}/", false);
-        TypeRestrictions typeRestrictions = new DataTypeRestrictions(Collections.singletonList(
+        TypeRestrictions typeRestrictions = new TypeRestrictions(Collections.singletonList(
             IsOfTypeConstraint.Types.STRING
         ));
         FieldSpec fieldSpecInSetWithTypedStringRestrictionsAndNullNotDisallowedd = FieldSpec.Empty
@@ -120,7 +120,7 @@ public class StandardFieldValueSourceEvaluatorTests {
             min = new DateTimeLimit(OffsetDateTime.MIN, false);
             max = new DateTimeLimit(OffsetDateTime.MAX, false);
         }};
-        TypeRestrictions typeRestrictions = new DataTypeRestrictions(Collections.singletonList(
+        TypeRestrictions typeRestrictions = new TypeRestrictions(Collections.singletonList(
             IsOfTypeConstraint.Types.DATETIME
         ));
         FieldSpec fieldSpecInSetWithTypedDateTimeRestrictionsAndNullNotDisallowed = FieldSpec.Empty
@@ -140,7 +140,7 @@ public class StandardFieldValueSourceEvaluatorTests {
                 max = new NumericLimit<>(new BigDecimal("1E+18"), false);
             }}
         ).withTypeRestrictions(
-            new DataTypeRestrictions(
+            new TypeRestrictions(
                 Collections.singletonList(IsOfTypeConstraint.Types.NUMERIC)
             )
         ).withNotNull(
@@ -174,7 +174,7 @@ public class StandardFieldValueSourceEvaluatorTests {
                 max = new NumericLimit<>(new BigDecimal("15.00000000000000000010"), false);
             }}
         ).withTypeRestrictions(
-            new DataTypeRestrictions(
+            new TypeRestrictions(
                 Collections.singletonList(IsOfTypeConstraint.Types.NUMERIC)
             )
         ).withNotNull();
@@ -210,7 +210,7 @@ public class StandardFieldValueSourceEvaluatorTests {
         FieldSpec fieldSpec = FieldSpec.Empty.withNumericRestrictions(
             restrictions
         ).withTypeRestrictions(
-            new DataTypeRestrictions(
+            new TypeRestrictions(
                 Collections.singletonList(IsOfTypeConstraint.Types.NUMERIC)
             )
         ).withNotNull();
@@ -239,7 +239,7 @@ public class StandardFieldValueSourceEvaluatorTests {
         FieldSpec fieldSpec = FieldSpec.Empty.withNumericRestrictions(
             new NumericRestrictions()
         ).withTypeRestrictions(
-            new DataTypeRestrictions(
+            new TypeRestrictions(
                 Collections.singletonList(IsOfTypeConstraint.Types.NUMERIC)
             )
         ).withNotNull();
@@ -271,7 +271,7 @@ public class StandardFieldValueSourceEvaluatorTests {
                 max = new NumericLimit<>(new BigDecimal("150.5"), false);
             }}
         ).withTypeRestrictions(
-            new DataTypeRestrictions(
+            new TypeRestrictions(
                 Collections.singletonList(IsOfTypeConstraint.Types.NUMERIC)
             )
         ).withNotNull();
@@ -298,7 +298,7 @@ public class StandardFieldValueSourceEvaluatorTests {
                 max = new NumericLimit<>(new BigDecimal("1.5E-20"), false);
             }}
         ).withTypeRestrictions(
-            new DataTypeRestrictions(
+            new TypeRestrictions(
                 Collections.singletonList(IsOfTypeConstraint.Types.NUMERIC)
             )
         ).withNotNull();
@@ -325,7 +325,7 @@ public class StandardFieldValueSourceEvaluatorTests {
                 max = new NumericLimit<>(new BigDecimal("3E-20"), false);
             }}
         ).withTypeRestrictions(
-            new DataTypeRestrictions(
+            new TypeRestrictions(
                 Collections.singletonList(IsOfTypeConstraint.Types.NUMERIC)
             )
         ).withNotNull();
